CollectDataService/requirements.txt
apted==1.0.3
asgiref==3.5.0
backports.zoneinfo==0.2.1
bleach==4.1.0
certifi==2021.10.8
charset-normalizer==2.0.9
commonmark==0.9.1
cssselect==1.1.0
dj-database-url==0.5.0
Django==3.2.12
django-cors-headers==3.11.0
djangorestframework==3.13.1
docutils==0.18.1
gunicorn==20.1.0
html-matcher==0.0.8
idna==3.3
importlib-metadata==4.11.3
keyring==23.5.0
lxml==4.8.0
packaging==21.3
parsel==1.6.0
pkginfo==1.8.2
psycopg2==2.9.3
Pygments==2.11.2
pyparsing==3.0.7
python-decouple==3.4
pytz==2021.3
readme-renderer==34.0
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.2.0
six==1.16.0
spmf==1.4
sqlparse==0.4.2
twine==4.0.0
typing_extensions==4.1.1
urllib3==1.26.7
w3lib==1.22.0
webencodings==0.5.1
XlsxWriter==3.0.3
zipp==3.8.0
